refactor: components 폴더 정리 및 쓰지 않는 컴포넌트 삭제
