+++ src/app/components/detail/Detail.tsx
