+++ src/app/components/layout/Background.tsx
