+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 // import Head from 'next/head';
 import './globals.css';
-import Header from './components/Header';
-import Footer from './components/Footer';
+import Header from './components/layout/Header';
+import Footer from './components/layout/Footer';
 
 export default function RootLayout({ children }: { children: React.ReactNode }) {
 	return (
