+++ src/app/user/layout.tsx
@@ -1,13 +1,11 @@
 // app/user/page.tsx
 import React from 'react';
-import Background from '@components/shared/Background';
+import Background from '@components/layout/Background';
 
 const Layout = ({ children }: { children: React.ReactNode }) => {
 	return (
 		<>
-			<div className="z-20 absolute left-1/2 -translate-x-1/2 max-w-inner mx-auto top-[7rem] w-full flex gap-7 flex-col">
-        {children}
-			</div>
+			<div className="z-20 absolute left-1/2 -translate-x-1/2 max-w-inner mx-auto top-[7rem] w-full flex gap-7 flex-col">{children}</div>
 			<Background />
 		</>
 	);
